simplify and clean up video predictor
