+++ sam3/perflib/__init__.py
@@ -4,5 +4,5 @@ import os
 
 is_enabled = False
 if os.getenv("USE_PERFLIB", "1") == "1":
-    print("Enabled the use of perflib.")
+    print("Enabled the use of perflib.\n", end="")
     is_enabled = True
